rework byte size from usize->u64 for buffer metrics

+++ lib/vector-core/buffers/src/config.rs
@@ -42,7 +42,7 @@ impl BufferTypeVisitor {
     {
         let mut kind: Option<BufferTypeKind> = None;
         let mut max_events: Option<usize> = None;
-        let mut max_size: Option<usize> = None;
+        let mut max_size: Option<u64> = None;
         let mut when_full: Option<WhenFull> = None;
         while let Some(key) = map.next_key::<String>()? {
             match key.as_str() {
@@ -237,14 +237,14 @@ pub enum BufferType {
     /// A buffer stage backed by an on-disk database, powered by LevelDB.
     #[serde(rename = "disk")]
     DiskV1 {
-        max_size: usize,
+        max_size: u64,
         #[serde(default)]
         when_full: WhenFull,
     },
     /// A buffer stage backed by disk.
     #[serde(rename = "disk_v2")]
     DiskV2 {
-        max_size: usize,
+        max_size: u64,
         #[serde(default)]
         when_full: WhenFull,
     },
