+++ lib/vector-core/buffers/src/variant/disk_v2.rs
@@ -14,16 +14,14 @@ use crate::disk_v2::{Buffer, DiskBufferConfig, Reader, Writer};
 use crate::topology::channel::{ReceiverAdapter, SenderAdapter};
 use crate::{topology::builder::IntoBuffer, Acker, Bufferable};
 
-//const MAX_BUFFERED_ITEMS: usize = 128;
-
 pub struct DiskV2Buffer {
     id: String,
     data_dir: PathBuf,
-    max_size: usize,
+    max_size: u64,
 }
 
 impl DiskV2Buffer {
-    pub fn new(id: String, data_dir: PathBuf, max_size: usize) -> Self {
+    pub fn new(id: String, data_dir: PathBuf, max_size: u64) -> Self {
         Self {
             id,
             data_dir,
